Add files via upload
Consider the maximum capacity of a network is 128 packets. The user starts his packets transmission by 1 packet. Then he/she increases the number of packets according to the following rules: If the number of transmitted packets < 64 packets, the number will be doubled. If the number of transmitted packets ≥ 64 packets, the number will be increased by 1. If the number of transmitted packets = 128 packets, the user will start over and transmit 1 packets and repeat the previous rules. The input is the size of the transmitted file. The output is the number of Transmissions used to transmit this file.